Export: Add Tags Export Plugin to export CSV file of tags for each image
This commit creates a new plugin 'TagsExportPlugin' created based on
the recommendations of Peter Carini/Joshua at Special Collections.

The plugin outputs a tab-delimited CSV file. At the top is a header
consisting of information/URLs for the project and metadata about the
date of export, version of the software, etc. All of these header
lines start with a hash (#).

After the header, there is a label line (Image Name, Tags), and then
one line for each image (with the data following the column labels).