Config: Set version values for software, csv format, and metadata format
This commit creates variables in the Yii "application-level
parameters" array to hold information about the version of the mg
software, the format of the csv export, and the format of the metadata
when embedded into exported images.

If there are other configuration values, we can place them here (or we
could move these configuration values elsewhere if there is a more
appropriate place for them to live).