PHP Metadata Toolkit: Make several tweaks to get XMP to append properly
This commit makes several tweaks to the pieces of the XMPAppend class
to make the code properly operate on an image and append new metadata
to the 'dc:description' field.

Several of these changes are temporary and are marked as such -- we'd
like to further understand the details of why these tweaks are
necessary before we make them wholly a part of the library.

In particular, I'm curious about the bits of the template XMP array
that were preventing the XMP data from being recognized by external
programs. I need to ask Evan Hunter (the original author of the code)
about that particular structure.