Export: Clean up Images Export Plugin, add correct system URL
This commit cleans up the Images Export Plugin a bit, removing
debugging lines and clarifying a few awkward lines w.r.t. directory
structure.

We also add the correct incantation (i.e. function call) to get the
URL for the installed system so that we can include that information
in our exported files.