tools: Create tools directory and add scripts for Markdown 1.0.1
This commit creates a new top-level tools/ directory to hold various
scripts and external code useful for working on or with the metadata
games project.

The first tool we're including is the perl code for markdown
1.0.1. We're using markdown syntax for our general documentation of
the mg project, so by including these tools in-tree anyone can tweak
and then re-generate the HTML output from the source text files.

The markdown scripts/docs are provided under the 3-clause BSD license
and are available from: http://daringfireball.net/projects/markdown/