Export: Fix non-static function calls
This commit fixes some function calls to methods on XMPAppend. We
instatiate an instance of XMPAppend so that we can make calls
dynamically instead of statically.

At the heart of it, none of these functions carry state; it would make
some sense to just refactor XMPAppend (and the libraries below it) so
that we could make all of our calls statically, however because this
refactoring might reach down deeper than the XMPAppend layer, this
would take a bit of time and would make it harder to upgrade the PHP
Metadata Toolkit library in the future.