Guess What: Add 'hints left' to score div; limit # of hints per config
This commit adds a 'hints left' box to the LHS score area of the Guess
What game interface. It acts (and currently looks) just like the
'guesses left' box above it.

This commit tweaks some of the pieces of the javascript behind hint
passing/showing so that we stop the hint countdown at zero, and don't
let the player see anymore hints.

This commit also does some cleanup and refactoring to support the
above changes.  (more similar cleanup is possible -- if we need to do
more surgery in the future).