Guess What: Clean up evaluateGuess; fix calls to areHintsAllowedLeft()
This commit cleans up some of the code in evaluateGuess to make it
more readable, and fixes some calls to areHintsAllowedLeft() so that
the function is properly called.

The behavior of the code seems to be correct from the standpoint of
the 'guesser', however even after we've run out of hints there's still
a message propagating back to the 'describer' that is triggering the
event for initiating the hint-giving mode. That message chain might be
important for updating the display of the 'describer' to let them know
that the 'guesser' is continuing to interact with the game (through
guesses), however we need to nip at least part of it in the bud to
prevent extra hints (and their fun but annoying negative-numbered
indicies!).