Guess What: Hide sendHintFormContainer and use new text when hints are gone
This commit hides the sendHintFormContainer div and uses a new
template of text when the hints are gone. Hiding the div (replacing it
with new text) and updating the #partner-waiting message both prevent
the 'Describer' from submitting new hints and also let that player
know that they need to wait for the other player to finish guessing
the image.