Guess What: Factor-out hint form message; fix typo in hint template
This commit separates changing text message from the HTML form. It
appears that JQuery does not directly support tweaking the HTML 'form'
elements dynamically for this use case.

We also fix a typo in the hint template that was preventing the proper
message from appearing the first time a hint was requested.