ZenTag: Add code to implement PASS button
This commit adds code to implement the PASS button for Zen Tag.

The basic idea is that when the user clicks the PASS button, we send
"__PASSONTHISTURN__" as the tag. The internals of the tag backend (in
MGTags.php) clean that up to "PASSONTHISTURN" when we handle it in
other code, though -- so be wary of that.

This commit implements the little pieces necessary to ignore this tag
in scoring, when selecting tags for single-player GuessWhat, etc. The
tag does show up in the backend when selecting all tags, but as that's
just the ADMIN interface, it's fine for it to be visible there.