ZenTagPlayOnce: Copy js from ZenTag into new file in preparation for split
This commit copies the code from mg.game.zentag.js into a new file
mg.game.zentagplayoncemoveon.js in preparation for ZenTagPlayOnce
splitting away from the ZenTag javascript file.

We need this split so that we can implement some slightly different
behavior in the two versions of the "zen tag" format. It's possible
that we'll be able to refactor these files to share a lot of code,
however we'll need to punt on that possibility for the time being.