Tools: Update fix-perms script to cover all files listed in the installer
This commit updates the fix-perms.sh script to include all of the
files and directories mentioned in the installer (and included in the
installer's Requirements Test).

This list is not directly linked to the code of the installer Test, so
if the directories and/or perms required for proper execution of MG
change, the fix-perms script will require corresponding updates.