docs: Update install/config information
This commit updates some of the install/config information to be more
accurate and to provide more help to those who are installing MG for
the first time.

There's still a lot more we can do to improve the documentation,
especially covering those q's and common pitfalls that we see when
others are testing out the software.