UserCounter: Tweak time variables that result in occasional error
We've sporadically seen an error on some systems (but not all) where
the user counter code has an issue with getting some variables set
from the DB properly.

We don't have a ton of time to dig through the code and figure out
exactly what's going wrong here, especially as all of the comments are
in German, so for now we're just going to not set the value
$this->user_time. The user-counter code isn't critical to game play or
other components, so we'll probably punt on this a bit longer.

Interestingly, the user counter seems to work correctly even with this
variable commented-out. Hmmm...