Export: Fix bug in timerange
This commit fixes the bug in the timerange export. We were casting a
string value as an int, resulting in only using the year component:

 "2013-04-01" ---(cast as int)--> 2013

We cast these values as a string instead, and the date ranges appear
to be interpreted properly. We could further sanitize this input, but
as we're going through the administrator interface, we largely assume
proper use.