NexTag: scale-down footer and move to sidepanel
This commit "moves" the footer to the sidepanel, essentially by
removing it (in javascript) from the footer area and creating a
similar footer in the sidepanel.

As long as we don't make too many game-specific changes to the footer,
we may want to combine these pieces of code in the future.