NexTag: Update layout of game interface per May 2 mockup
This commit updates the layout of the NexTag interface to match the
May 2 version of the game mockup. The image is reduced to fit the
window size and the text box and button are centered, moved to the
bottom of the screen, and offset in a dark box, more closely matching
the original layout style of ZenTag.

We're still having issues with getting max-height: to work
properly. It appears that we need to specify a hard height on the
parent of the image in order to use max-height: in CSS (percentages do
not seem to work). I might need to look to jQuery to get the scaling
we desire.