NexTag: Add placeholder text to the text input box
This commit adds placeholder text to the text input box (similar to
what we had in ZenTag a couple of iterations ago), per request by
Sukie/Mary. The placeholder attribute seems to work most places, but
if we have an extensive IE installation, we can re-work this with a
bit of javascript if IE hasn't implemented this feature yet.