NexTag: Tweak review screen to show images in table, resized to window
This commit tweaks the final review screen to show 4 images in a
table, with the height of the table set by javascript to fit within
the bounds of the screen.

For now, we're left- and right- aligning the images so that they butt
up against each other in the center of the table, but if we do add a
border to the table then perhaps we can center them in each <td>
without it looking awkward.