EasyTag: Stub-in external "mobile" HTML5 game
This commit stubs-in an HTML5 game that is "external" to the Yii
framework. We've based it on the 'Mobile' game and are using the
Mobile name and entrance points into the code for now, although we
might abstract-away those pieces even further for our actual
implementation of various mobile games.