NexTag: Revamp logos for sidepanel
This commit revamps the logos used in the sidepanel in NexTag. We add
a logo for tiltfactor, reuse the metadatagames logo, and return to the
larger (and wider) NEH logo in a 3-row stack.

We'll definitely want to tweak the display of these images -- do we
want to use a white BG throughout, or have some choice/variety?