TagsByGame: Add new Export plugin to provide tags grouped by game/image
This export plugin provides tags grouped by game and image. It
provides a number of pieces of information including Game, Image
Set(s), Image, Image ID, and Tags.

This plugin is currently a work in progress and may change or be
tweaked in the coming weeks. Although the output has been tested,
please do not make any assumptions about the data columns or their
position in the file until we've firmed-up our plans for this plugin.