Export: Refactor and implement code to handle Active status for each plugin
This commit refactors code handling the 'Active' status of an export
plugin so that it is implemented consistently across all of these
plugins.

The value $activeByDefault is configured in MGExportPlugin to FALSE,
but may be set on a per-plugin basis to TRUE (e.g. the Tags Export and
Game Statistics plugins are currently both Active by default)