Yii 1.1.13 upgrade: Fix bug in batch actions
This commit fixes a bug we discovered after the upgrade to Yii v1.1.13
where running certain batch actions did not seem to have any
effect. It also fixes the bug where the 'Delete' buttons on the
admin/imageSet page were not effective.

Credit to Nikolay Kondikov <nikolay.kondikov@sirma.bg> for the fix.